+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file is used to list changes made in each version of the openstack-dashboar
 * Update local_settings from 0644 to 0640
 * Fix python-ibm-db-django package polluting common package attribute
 * Allow some ceitificate options to be configured
+* Add sensitive flag to private key and certificate file resources
 
 ## 9.1
 * python_packages database client attributes have been moved to the -common cookbook